# Optimize pack_fp4_to_uint8 for fp4