Much faster impl now that uses torch to run on CPU or GPU. We can make it faster by adding `@torch.compile`

```
python benchmark_modelopt_quant.py      
Creating tensor of shape (8192, 8192) on cpu...
Benchmarking on tensor of shape torch.Size([8192, 8192]) (67,108,864 elements) on cpu
Iter 1/3 - Pack: 1.8838s
Iter 2/3 - Pack: 1.9518s
Iter 3/3 - Pack: 1.9156s
Iter 1/3 - Unpack: 0.1986s
Iter 2/3 - Unpack: 0.2034s
Iter 3/3 - Unpack: 0.2029s

Benchmark Results:
Device: cpu
Tensor shape: 8192x8192 (67,108,864 elements)
Average pack time: 1.9171s (35.01M elements/s)
Average unpack time: 0.2016s (332.83M elements/s)
Compression ratio: 8.00x
```

```
python benchmark_modelopt_quant.py --gpu
Creating tensor of shape (8192, 8192) on cuda...
Benchmarking on tensor of shape torch.Size([8192, 8192]) (67,108,864 elements) on cuda
Iter 1/3 - Pack: 0.0162s
Iter 2/3 - Pack: 0.0200s
Iter 3/3 - Pack: 0.0201s
Iter 1/3 - Unpack: 0.0192s
Iter 2/3 - Unpack: 0.0036s
Iter 3/3 - Unpack: 0.0038s

Benchmark Results:
Device: cuda
Tensor shape: 8192x8192 (67,108,864 elements)
Average pack time: 0.0188s (3576.69M elements/s)
Average unpack time: 0.0089s (7565.92M elements/s)
Compression ratio: 8.00x
```